Moving tests back to their own directory
