+++ src/canhydro/CylinderCollection.py
@@ -6,6 +6,7 @@ from __future__ import annotations
 import copy
 import math
 import sys
+from itertools import chain
 
 import matplotlib.pyplot as plt
 import networkx as nx
@@ -19,9 +20,8 @@ from canhydro.Cylinder import create_cyl
 from canhydro.DataClasses import Flow
 from canhydro.geometry import concave_hull  # ,vectorized_get_projection)
 from canhydro.geometry import draw_cyls, furthest_point, get_projected_overlap
-from canhydro.global_vars import log
+from canhydro.global_vars import config_vars, log
 from canhydro.utils import intermitent_log, lam_filter, save_file
-from canhydro.global_vars import config_vars
 
 sys.stdout = LogFile()
 
@@ -75,7 +75,7 @@ class CylinderCollection:
 
         # Graph and Attributes
         self.graph = None
-        self.graph = None
+        self.digraph = None
         self.min_graph = None
         self.flows = None
         self.drip_points = {"x": np.nan, "y": np.nan, "z": np.nan, "flow_id": np.nan}
@@ -207,7 +207,7 @@ class CylinderCollection:
     def __eq__(self, other):
         if len(self.cylinders) == 0 or len(other.cylinders) == 0:
             raise AttributeError(
-                "One or both Cylinder Collections contain no Cyliders (Did you forget to initialize?)"
+                "One or both Cylinder Collections contain no Cylinders (Did you forget to initialize?)"
             )
         if type(self) == type(other):
             raise TypeError(
@@ -346,6 +346,22 @@ class CylinderCollection:
         gr.add_edges_from(edges)
         self.graph = gr
 
+    def initialize_digraph_from(
+        self, in_flow_grade_lim=config_vars["in_flow_grade_lim"]
+    ):
+        """This function initialized edge attributes as cylinder objects"""
+        gr = nx.DiGraph()
+        edges = (
+            (
+                (int(cyl.cyl_id), int(cyl.parent_id), {"cylinder": cyl})
+                if cyl.angle >= in_flow_grade_lim
+                else (int(cyl.parent_id), int(cyl.cyl_id), {"cylinder": cyl})
+            )
+            for cyl in self.cylinders
+        )
+        gr.add_edges_from(edges)
+        self.digraph = gr
+
     def sum_over_graph(self):
         """This function initialized edge attributes as FULL cylinder dicts"""
         gr = self.graph
@@ -382,19 +398,64 @@ class CylinderCollection:
         """
         trunk_nodes = self.get_trunk_nodes()
 
-        trunk_contraction, titans = self.contracted_nodes(
-            self.trunk_nodes, 0, self.graph
-        )
+        trunk_contraction, titans = self.contracted_nodes(trunk_nodes, 0, self.graph)
         trunk_paths = nx.shortest_path(trunk_contraction, target=0)
         dists = {node: len(path) - 1 for node, path in trunk_paths.items()}
         return dists
 
-    def traverse_until(self, start_node, traversed):
+    def traverse_until(
+        self, graph: nx.Graph, start_node: int, to_return: list[tuple] = []
+    ):
         """
         Intended to take a starting point and recursively
         travel adjacent edges until no new edge is found or until
         a node in the stop set is reached.
         """
+        neighbors = nx.neighbors(graph, start_node)
+
+    def find_flow_components_digraph(self, inFlowGradeLim=-1 / 6):
+        g = self.graph
+        if type(g) == nx.Graph:
+            msg = "Find Flow Digraph invoked for undirected graph"
+            log.error(msg)
+            raise TypeError(msg)
+        root_node = 0
+
+        drip_edges = [
+            (u, v)
+            for u, v, attr in g.edges(data=True)
+            if attr["cylinder"].angle < inFlowGradeLim
+        ]
+        in_flow_graph = copy.deepcopy(g)  # This could probably just be a subgraph...
+        in_flow_graph.remove_edges_from(drip_edges)
+
+        stem_flow_component = g.subgraph(g.predecessors(root_node)).copy()
+
+        G_drip = copy.deepcopy(g)
+        G_drip.remove_edges_from(in_flow_graph.edges())
+        out_flow_components = nx.weakly_connected_components(G_drip)
+        drip_flow_components = []
+        for out_flow_component in out_flow_components:
+            dc_nodes = out_flow_component.nodes
+            predecessors = [g.predecessors(node) for node in dc_nodes]
+            drip_flow_nodes = chain(*predecessors)
+            drip_flow_component = g.subgraph(drip_flow_nodes).copy()
+            drip_flow_components.append(drip_flow_component)
+
+        log.info(
+            f"{self.file_name} found to have {len(drip_flow_components)} drip components"
+        )
+        stem_edge_ids = [
+            max([u, v])
+            for (u, v) in stem_flow_component.edges()
+            if G_drip.degree(max([u, v])) == 0
+        ]
+        for cyl in self.cylinders:
+            if cyl.cyl_id in stem_edge_ids:
+                cyl.is_stem = True
+        self.stem_flow_component = stem_flow_component
+        self.drip_flow_components = drip_flow_components
+        self.drip_graph = G_drip
 
     # @profile
     def find_flow_components(self, inFlowGradeLim=-1 / 6):
@@ -422,8 +483,13 @@ class CylinderCollection:
         G_drip.remove_edges_from(in_flow_graph.edges())
         drip_flow_components = nx.connected_components(G_drip)
         drip_edge_graphs = [g.subgraph(c).copy() for c in drip_flow_components]
-        tributaries = []
-        confluences = [max(g.nodes()) for g in out_flow_graphs]
+
+        for drip_graph in drip_edge_graphs:
+            dg_nodes = drip_graph.nodes()
+            connected_in_flows = [
+                g.subgraph(nx.node_connected_component(in_flow_graph, node)).copy()
+                for node in dg_nodes
+            ]
 
         breakpoint()
         component_graphs = [
@@ -466,7 +532,7 @@ class CylinderCollection:
             edge_attributes[(u, v)] = {"dripNode": 0, "flowType": "stem", "flowID": 0}
         # log.info("attempting to sum stem edges ")
 
-        # this probably doesnt belong here but its efficient to do it now
+        # this probably doesn't belong here but its efficient to do it now
         # is 'needed' for statistics section
         stem_polys = [cyl.projected_data[plane]["polygon"] for _, _, cyl in stem_edges]
         self.stem_polys[plane] = stem_polys
@@ -792,7 +858,7 @@ class CylinderCollection:
         Returns a plot showing the locations of the drip points, subject
         to input params
 
-        a_lambda: funtion to filter drip points displayed (e.g. those with projected area>10m^2 )
+        a_lambda: function to filter drip points displayed (e.g. those with projected area>10m^2 )
         scale: how large of a boundary to draw around drip points
         """
         drip_point_locs = self.get_drip_points()
