--- src/canhydro/tests/expected_results.py
@@ -1,2205 +1,0 @@
-from __future__ import annotations
-
-from canhydro.DataClasses import Flow
-
-ten_cyls_cyls = [
-    "Cylinder(cyl_id=0.0, x=array([0.076822, 0.076822]), y=array([-3.386935, -3.386935]), z=array([-0.595933, -0.5315  ]), radius=0.222504, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=-1.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0, dy=0.0, dz=0.06443300000000007, surface_area=0.09007951593260982, sa_to_vol=8.98907453673384, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=1.0, x=array([0.076822, 0.058561]), y=array([-3.386935, -3.328717]), z=array([-0.5315  , -0.085599]), radius=0.211908, length=0.450057, branch_order=0.0, branch_id=0.0, volume=0.063491, parent_id=0.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.018261, dy=0.05821799999999966, dz=0.445901, surface_area=0.5992316474954436, sa_to_vol=9.438056535500206, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=2.0, x=array([0.058561, 0.062703]), y=array([-3.328717, -3.334542]), z=array([-0.085599,  0.232646]), radius=0.197783, length=0.318325, branch_order=0.0, branch_id=0.0, volume=0.03912, parent_id=1.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.004141999999999993, dy=-0.005824999999999747, dz=0.318245, surface_area=0.3955847820488214, sa_to_vol=10.112085430695844, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=3.0, x=array([0.062703, 0.06246 ]), y=array([-3.334542, -3.357087]), z=array([0.232646, 0.499611]), radius=0.198531, length=0.267916, branch_order=0.0, branch_id=0.0, volume=0.033174, parent_id=2.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.00024299999999999322, dy=-0.022545000000000037, dz=0.266965, surface_area=0.33420031048164517, sa_to_vol=10.07416381749699, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=4.0, x=array([0.06246 , 0.073532]), y=array([-3.357087, -3.374155]), z=array([0.499611, 0.779853]), radius=0.188766, length=0.280979, branch_order=0.0, branch_id=0.0, volume=0.031454, parent_id=3.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.011071999999999999, dy=-0.017068000000000083, dz=0.280242, surface_area=0.33325563682540077, sa_to_vol=10.595016113225686, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=5.0, x=array([0.073532, 0.067962]), y=array([-3.374155, -3.393301]), z=array([0.779853, 1.05038 ]), radius=0.187445, length=0.27126, branch_order=0.0, branch_id=0.0, volume=0.029942, parent_id=4.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0055700000000000055, dy=-0.019146000000000107, dz=0.2705270000000001, surface_area=0.3194769179782343, sa_to_vol=10.669858993328244, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=6.0, x=array([0.067962, 0.062838]), y=array([-3.393301, -3.386649]), z=array([1.05038 , 1.311286]), radius=0.19627, length=0.261041, branch_order=0.0, branch_id=0.0, volume=0.031591, parent_id=5.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.00512399999999999, dy=0.006652000000000324, dz=0.26090599999999986, surface_area=0.3219159648746658, sa_to_vol=10.190116326633085, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=7.0, x=array([0.062838, 0.055451]), y=array([-3.386649, -3.429066]), z=array([1.311286, 1.581252]), radius=0.182718, length=0.273378, branch_order=1.0, branch_id=0.0, volume=0.028673, parent_id=6.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.007387000000000005, dy=-0.04241700000000037, dz=0.26996600000000015, surface_area=0.31385190075534425, sa_to_vol=10.94590383829192, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=8.0, x=array([0.055451, 0.073181]), y=array([-3.429066, -3.465325]), z=array([1.581252, 1.802064]), radius=0.203862, length=0.22447, branch_order=1.0, branch_id=0.0, volume=0.029308, parent_id=7.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.017729999999999996, dy=-0.03625899999999982, dz=0.220812, surface_area=0.2875242342525162, sa_to_vol=9.810435179900239, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=9.0, x=array([0.073181, 0.062266]), y=array([-3.465325, -3.478196]), z=array([1.802064, 2.056271]), radius=0.196833, length=0.254767, branch_order=1.0, branch_id=0.0, volume=0.031009, parent_id=8.0, reverse_branch_order=31.0, segment_id=1.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.010914999999999994, dy=-0.012871000000000077, dz=0.25420700000000007, surface_area=0.315080084456099, sa_to_vol=10.160923746528395, slope=0.0, is_stem=False, is_divide=False)",
-]
-
-ez_projection_cyls = [
-    "Cylinder(cyl_id=0.0, x=array([1., 4.]), y=array([1., 6.]), z=array([1., 7.]), radius=1.0, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=-1.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=3.0, dy=5.0, dz=6.0, surface_area=0.40484447889750186, sa_to_vol=40.399608711456125, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=1.0, x=array([4., 7.]), y=array([ 6., 11.]), z=array([7., 1.]), radius=1.0, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=0.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=3.0, dy=5.0, dz=-6.0, surface_area=0.40484447889750186, sa_to_vol=40.399608711456125, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=1.0, x=array([ 7., 10.]), y=array([11., 16.]), z=array([ 1., -5.]), radius=1.0, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=0.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=3.0, dy=5.0, dz=-6.0, surface_area=0.40484447889750186, sa_to_vol=40.399608711456125, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=1.0, x=array([10.,  7.]), y=array([16., 21.]), z=array([ -5., -11.]), radius=1.0, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=0.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-3.0, dy=5.0, dz=-6.0, surface_area=0.40484447889750186, sa_to_vol=40.399608711456125, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=1.0, x=array([ 7., 10.]), y=array([21., 16.]), z=array([-11., -17.]), radius=1.0, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=0.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=3.0, dy=-5.0, dz=-6.0, surface_area=0.40484447889750186, sa_to_vol=40.399608711456125, slope=0.0, is_stem=False, is_divide=False)",
-]
-
-happy_path_cyls = [
-    "Cylinder(cyl_id=0.0, x=array([0.076822, 0.076822]), y=array([-3.386935, -3.386935]), z=array([-0.595933, -0.5315  ]), radius=0.222504, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=-1.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0, dy=0.0, dz=0.06443300000000007, surface_area=0.09007951593260982, sa_to_vol=8.98907453673384, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=1.0, x=array([0.076822, 0.058561]), y=array([-3.386935, -3.328717]), z=array([-0.5315  , -0.085599]), radius=0.211908, length=0.450057, branch_order=0.0, branch_id=0.0, volume=0.063491, parent_id=0.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.018261, dy=0.05821799999999966, dz=0.445901, surface_area=0.5992316474954436, sa_to_vol=9.438056535500206, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=2.0, x=array([0.058561, 0.062703]), y=array([-3.328717, -3.334542]), z=array([-0.085599,  0.232646]), radius=0.197783, length=0.318325, branch_order=0.0, branch_id=0.0, volume=0.03912, parent_id=1.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.004141999999999993, dy=-0.005824999999999747, dz=0.318245, surface_area=0.3955847820488214, sa_to_vol=10.112085430695844, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=3.0, x=array([0.062703, 0.06246 ]), y=array([-3.334542, -3.357087]), z=array([0.232646, 0.499611]), radius=0.198531, length=0.267916, branch_order=0.0, branch_id=0.0, volume=0.033174, parent_id=2.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.00024299999999999322, dy=-0.022545000000000037, dz=0.266965, surface_area=0.33420031048164517, sa_to_vol=10.07416381749699, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=4.0, x=array([0.06246 , 0.073532]), y=array([-3.357087, -3.374155]), z=array([0.499611, 0.779853]), radius=0.188766, length=0.280979, branch_order=0.0, branch_id=0.0, volume=0.031454, parent_id=3.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.011071999999999999, dy=-0.017068000000000083, dz=0.280242, surface_area=0.33325563682540077, sa_to_vol=10.595016113225686, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=5.0, x=array([0.073532, 0.067962]), y=array([-3.374155, -3.393301]), z=array([0.779853, 1.05038 ]), radius=0.187445, length=0.27126, branch_order=0.0, branch_id=0.0, volume=0.029942, parent_id=4.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0055700000000000055, dy=-0.019146000000000107, dz=0.2705270000000001, surface_area=0.3194769179782343, sa_to_vol=10.669858993328244, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=6.0, x=array([0.067962, 0.062838]), y=array([-3.393301, -3.386649]), z=array([1.05038 , 1.311286]), radius=0.19627, length=0.261041, branch_order=0.0, branch_id=0.0, volume=0.031591, parent_id=5.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.00512399999999999, dy=0.006652000000000324, dz=0.26090599999999986, surface_area=0.3219159648746658, sa_to_vol=10.190116326633085, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=7.0, x=array([0.062838, 0.055451]), y=array([-3.386649, -3.429066]), z=array([1.311286, 1.581252]), radius=0.182718, length=0.273378, branch_order=0.0, branch_id=0.0, volume=0.028673, parent_id=6.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.007387000000000005, dy=-0.04241700000000037, dz=0.26996600000000015, surface_area=0.31385190075534425, sa_to_vol=10.94590383829192, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=8.0, x=array([0.055451, 0.073181]), y=array([-3.429066, -3.465325]), z=array([1.581252, 1.802064]), radius=0.203862, length=0.22447, branch_order=0.0, branch_id=0.0, volume=0.029308, parent_id=7.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.017729999999999996, dy=-0.03625899999999982, dz=0.220812, surface_area=0.2875242342525162, sa_to_vol=9.810435179900239, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=9.0, x=array([0.073181, 0.062266]), y=array([-3.465325, -3.478196]), z=array([1.802064, 2.056271]), radius=0.196833, length=0.254767, branch_order=0.0, branch_id=0.0, volume=0.031009, parent_id=8.0, reverse_branch_order=31.0, segment_id=1.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.010914999999999994, dy=-0.012871000000000077, dz=0.25420700000000007, surface_area=0.315080084456099, sa_to_vol=10.160923746528395, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=10.0, x=array([0.062266, 0.027591]), y=array([-3.478196, -3.518976]), z=array([2.056271, 2.315272]), radius=0.20932, length=0.264475, branch_order=0.0, branch_id=0.0, volume=0.036404, parent_id=9.0, reverse_branch_order=30.0, segment_id=2.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.034675, dy=-0.040779999999999816, dz=0.25900100000000004, surface_area=0.34783655426922827, sa_to_vol=9.554899304176143, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=11.0, x=array([ 0.027591, -0.001305]), y=array([-3.518976, -3.560212]), z=array([2.315272, 2.468256]), radius=0.20932, length=0.161058, branch_order=0.0, branch_id=0.0, volume=0.022169, parent_id=10.0, reverse_branch_order=30.0, segment_id=2.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.028896, dy=-0.04123600000000005, dz=0.15298399999999956, surface_area=0.21182289349652467, sa_to_vol=9.554914226917075, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=12.0, x=array([-0.001305,  0.034777]), y=array([-3.560212, -3.438469]), z=array([2.468256, 2.858883]), radius=0.124765, length=0.410746, branch_order=0.0, branch_id=0.0, volume=0.020087, parent_id=11.0, reverse_branch_order=29.0, segment_id=3.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.036082, dy=0.12174299999999993, dz=0.3906270000000003, surface_area=0.3219926676132853, sa_to_vol=16.029903301303595, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=13.0, x=array([ 0.034777, -0.065825]), y=array([-3.438469, -3.331656]), z=array([2.858883, 3.06936 ]), radius=0.115601, length=0.256574, branch_order=0.0, branch_id=0.0, volume=0.010772, parent_id=12.0, reverse_branch_order=28.0, segment_id=4.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.100602, dy=0.10681299999999982, dz=0.21047700000000003, surface_area=0.18636060179968356, sa_to_vol=17.300464333427733, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=14.0, x=array([-0.065825, -0.058826]), y=array([-3.331656, -3.424373]), z=array([3.06936 , 3.270241]), radius=0.112207, length=0.221356, branch_order=0.0, branch_id=0.0, volume=0.008755, parent_id=13.0, reverse_branch_order=27.0, segment_id=5.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.006998999999999991, dy=-0.09271699999999994, dz=0.20088099999999987, surface_area=0.1560598257866162, sa_to_vol=17.8252228197163, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=15.0, x=array([-0.058826, -0.057322]), y=array([-3.424373, -3.428676]), z=array([3.270241, 3.422008]), radius=0.109626, length=0.151835, branch_order=0.0, branch_id=0.0, volume=0.005733, parent_id=14.0, reverse_branch_order=27.0, segment_id=5.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0015040000000000053, dy=-0.004302999999999724, dz=0.15176699999999999, surface_area=0.10458401973974013, sa_to_vol=18.242459399919785, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=16.0, x=array([-0.057322, -0.055536]), y=array([-3.428676, -3.418364]), z=array([3.422008, 3.567363]), radius=0.107314, length=0.145732, branch_order=0.0, branch_id=0.0, volume=0.005272, parent_id=15.0, reverse_branch_order=27.0, segment_id=5.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.001785999999999996, dy=0.010311999999999877, dz=0.1453549999999999, surface_area=0.09826326185150316, sa_to_vol=18.63870672448846, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=17.0, x=array([-0.055536, -0.056154]), y=array([-3.418364, -3.416518]), z=array([3.567363, 3.706142]), radius=0.119733, length=0.138792, branch_order=0.0, branch_id=0.0, volume=0.006251, parent_id=16.0, reverse_branch_order=27.0, segment_id=5.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0006180000000000005, dy=0.0018460000000000143, dz=0.13877899999999999, surface_area=0.10441386370516216, sa_to_vol=16.703545625525862, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=18.0, x=array([-0.056154, -0.058504]), y=array([-3.416518, -3.401558]), z=array([3.706142, 3.857456]), radius=0.108127, length=0.15207, branch_order=0.0, branch_id=0.0, volume=0.005585, parent_id=17.0, reverse_branch_order=27.0, segment_id=5.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.002349999999999998, dy=0.014959999999999862, dz=0.15131400000000017, surface_area=0.10331361735026955, sa_to_vol=18.49840955242069, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=19.0, x=array([-0.058504, -0.057836]), y=array([-3.401558, -3.297242]), z=array([3.857456, 4.002476]), radius=0.083356, length=0.178643, branch_order=0.0, branch_id=0.0, volume=0.0039, parent_id=18.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0006680000000000019, dy=0.1043160000000003, dz=0.1450199999999997, surface_area=0.09356269820285773, sa_to_vol=23.99043543663019, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=20.0, x=array([-0.057836, -0.059261]), y=array([-3.297242, -3.271659]), z=array([4.002476, 4.121282]), radius=0.078515, length=0.121537, branch_order=0.0, branch_id=0.0, volume=0.002354, parent_id=19.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0014250000000000027, dy=0.02558299999999969, dz=0.11880600000000019, surface_area=0.059957154767666984, sa_to_vol=25.470329128150798, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=21.0, x=array([-0.059261, -0.06571 ]), y=array([-3.271659, -3.257999]), z=array([4.121282, 4.217405]), radius=0.076817, length=0.097302, branch_order=0.0, branch_id=0.0, volume=0.001804, parent_id=20.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.006449000000000003, dy=0.013660000000000228, dz=0.0961230000000004, surface_area=0.04696334018155055, sa_to_vol=26.032893670482565, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=22.0, x=array([-0.06571 , -0.066685]), y=array([-3.257999, -3.235245]), z=array([4.217405, 4.322664]), radius=0.073089, length=0.107695, branch_order=0.0, branch_id=0.0, volume=0.001807, parent_id=21.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0009749999999999898, dy=0.02275399999999994, dz=0.10525899999999933, surface_area=0.04945696126104696, sa_to_vol=27.369652053706123, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=23.0, x=array([-0.066685, -0.078122]), y=array([-3.235245, -3.222873]), z=array([4.322664, 4.434789]), radius=0.076579, length=0.113385, branch_order=0.0, branch_id=0.0, volume=0.002089, parent_id=22.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.011437000000000003, dy=0.01237200000000005, dz=0.1121250000000007, surface_area=0.05455633200149195, sa_to_vol=26.116003830297725, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=24.0, x=array([-0.078122, -0.084576]), y=array([-3.222873, -3.199114]), z=array([4.434789, 4.537954]), radius=0.0772, length=0.106061, branch_order=0.0, branch_id=0.0, volume=0.001986, parent_id=23.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.006454000000000001, dy=0.023759000000000086, dz=0.10316499999999973, surface_area=0.05144615078196056, sa_to_vol=25.90440623462264, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=25.0, x=array([-0.084576, -0.094275]), y=array([-3.199114, -3.192158]), z=array([4.537954, 4.642884]), radius=0.077654, length=0.105607, branch_order=0.0, branch_id=0.0, volume=0.002001, parent_id=24.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.009699, dy=0.00695599999999974, dz=0.10492999999999952, surface_area=0.05152718362800013, sa_to_vol=25.750716455772174, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=26.0, x=array([-0.094275, -0.109448]), y=array([-3.192158, -3.177965]), z=array([4.642884, 4.742207]), radius=0.07615, length=0.101473, branch_order=0.0, branch_id=0.0, volume=0.001849, parent_id=25.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.015173000000000006, dy=0.014193000000000122, dz=0.09932300000000005, surface_area=0.048551234412734304, sa_to_vol=26.25810406313375, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=27.0, x=array([-0.109448, -0.121389]), y=array([-3.177965, -3.17529 ]), z=array([4.742207, 4.8489  ]), radius=0.075838, length=0.107392, branch_order=0.0, branch_id=0.0, volume=0.00194, parent_id=26.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.011940999999999993, dy=0.002674999999999983, dz=0.10669300000000081, surface_area=0.051172739833141495, sa_to_vol=26.377700944918296, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=28.0, x=array([-0.121389, -0.136877]), y=array([-3.17529 , -3.164554]), z=array([4.8489  , 4.958056]), radius=0.084646, length=0.110771, branch_order=0.0, branch_id=0.0, volume=0.002493, parent_id=27.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.015488000000000002, dy=0.010736000000000079, dz=0.10915599999999959, surface_area=0.058913169040474944, sa_to_vol=23.631435635970696, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=29.0, x=array([-0.136877, -0.131187]), y=array([-3.164554, -3.141393]), z=array([4.958056, 5.079266]), radius=0.075341, length=0.123534, branch_order=0.0, branch_id=0.0, volume=0.002203, parent_id=28.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.005690000000000001, dy=0.023160999999999987, dz=0.1212099999999996, surface_area=0.058478705801968595, sa_to_vol=26.54503213888724, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=30.0, x=array([-0.131187, -0.137226]), y=array([-3.141393, -3.123301]), z=array([5.079266, 5.179019]), radius=0.074351, length=0.10156, branch_order=0.0, branch_id=0.0, volume=0.001764, parent_id=29.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.006038999999999989, dy=0.018091999999999775, dz=0.09975300000000065, surface_area=0.047444882410218554, sa_to_vol=26.896191842527525, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=31.0, x=array([-0.137226, -0.136859]), y=array([-3.123301, -3.10119 ]), z=array([5.179019, 5.281189]), radius=0.072841, length=0.104536, branch_order=0.0, branch_id=0.0, volume=0.001742, parent_id=30.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.00036699999999997845, dy=0.022111000000000214, dz=0.1021700000000001, surface_area=0.04784335709638261, sa_to_vol=27.464613717785657, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=32.0, x=array([-0.136859, -0.140386]), y=array([-3.10119 , -3.087488]), z=array([5.281189, 5.378628]), radius=0.077031, length=0.098461, branch_order=0.0, branch_id=0.0, volume=0.001835, parent_id=31.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0035270000000000024, dy=0.013701999999999881, dz=0.09743899999999961, surface_area=0.04765512866679056, sa_to_vol=25.970097366098397, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=33.0, x=array([-0.140386, -0.149544]), y=array([-3.087488, -3.072912]), z=array([5.378628, 5.479159]), radius=0.084333, length=0.101995, branch_order=0.0, branch_id=0.0, volume=0.002279, parent_id=32.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.009158, dy=0.014575999999999922, dz=0.10053100000000015, surface_area=0.05404509698472581, sa_to_vol=23.714390954245637, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=34.0, x=array([-0.149544, -0.155951]), y=array([-3.072912, -3.064858]), z=array([5.479159, 5.581015]), radius=0.084304, length=0.102375, branch_order=0.0, branch_id=0.0, volume=0.002286, parent_id=33.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.006406999999999996, dy=0.008054000000000006, dz=0.10185599999999972, surface_area=0.05422779734222088, sa_to_vol=23.721696125205984, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=35.0, x=array([-0.155951, -0.163805]), y=array([-3.064858, -3.049039]), z=array([5.581015, 5.684423]), radius=0.078626, length=0.104905, branch_order=0.0, branch_id=0.0, volume=0.002037, parent_id=34.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.007854, dy=0.015819000000000027, dz=0.10340799999999994, surface_area=0.051825349371885314, sa_to_vol=25.441997727975114, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=36.0, x=array([-0.163805, -0.142775]), y=array([-3.049039, -3.034255]), z=array([5.684423, 5.77434 ]), radius=0.082411, length=0.09352, branch_order=0.0, branch_id=0.0, volume=0.001995, parent_id=35.0, reverse_branch_order=17.0, segment_id=6.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.021029999999999993, dy=0.01478400000000013, dz=0.0899169999999998, surface_area=0.04842499120840983, sa_to_vol=24.27317855058137, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=37.0, x=array([-0.142775, -0.183385]), y=array([-3.034255, -3.040343]), z=array([5.77434 , 5.908583]), radius=0.082411, length=0.140382, branch_order=0.0, branch_id=0.0, volume=0.002995, parent_id=36.0, reverse_branch_order=16.0, segment_id=7.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.04060999999999998, dy=-0.006088000000000093, dz=0.13424300000000056, surface_area=0.07269030277821846, sa_to_vol=24.27055184581585, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=38.0, x=array([-0.183385, -0.196242]), y=array([-3.040343, -3.022512]), z=array([5.908583, 6.02986 ]), radius=0.074488, length=0.123254, branch_order=0.0, branch_id=0.0, volume=0.002148, parent_id=37.0, reverse_branch_order=16.0, segment_id=7.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.012857000000000007, dy=0.017831000000000152, dz=0.12127700000000008, surface_area=0.05768557214524569, sa_to_vol=26.855480514546407, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=39.0, x=array([-0.196242, -0.185009]), y=array([-3.022512, -3.005524]), z=array([6.02986 , 6.139235]), radius=0.074488, length=0.111255, branch_order=0.0, branch_id=0.0, volume=0.001939, parent_id=38.0, reverse_branch_order=16.0, segment_id=7.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.011232999999999993, dy=0.016988000000000003, dz=0.109375, surface_area=0.052069777281218525, sa_to_vol=26.853933615894032, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=40.0, x=array([-0.185009, -0.247645]), y=array([-3.005524, -3.027011]), z=array([6.139235, 6.26064 ]), radius=0.074304, length=0.13829, branch_order=0.0, branch_id=0.0, volume=0.002399, parent_id=39.0, reverse_branch_order=15.0, segment_id=8.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.062636, dy=-0.021487000000000034, dz=0.12140500000000021, surface_area=0.06456287162923349, sa_to_vol=26.9124100163541, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=41.0, x=array([-0.247645, -0.255977]), y=array([-3.027011, -3.020758]), z=array([6.26064 , 6.381853]), radius=0.073989, length=0.121659, branch_order=0.0, branch_id=0.0, volume=0.002092, parent_id=40.0, reverse_branch_order=15.0, segment_id=8.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.008332000000000006, dy=0.006253000000000064, dz=0.12121300000000002, surface_area=0.05655763858872179, sa_to_vol=27.035200090211177, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=42.0, x=array([-0.255977, -0.180642]), y=array([-3.020758, -2.98338 ]), z=array([6.381853, 6.458277]), radius=0.05174, length=0.113636, branch_order=0.0, branch_id=0.0, volume=0.000956, parent_id=41.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.07533500000000001, dy=0.03737799999999991, dz=0.07642399999999938, surface_area=0.036942155397618956, sa_to_vol=38.642421964036565, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=43.0, x=array([-0.180642, -0.166534]), y=array([-2.98338 , -2.947926]), z=array([6.458277, 6.514814]), radius=0.051677, length=0.068208, branch_order=0.0, branch_id=0.0, volume=0.000572, parent_id=42.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.01410800000000001, dy=0.0354540000000001, dz=0.056537000000000504, surface_area=0.022146876166860903, sa_to_vol=38.71831497702955, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=44.0, x=array([-0.166534, -0.17596 ]), y=array([-2.947926, -2.948445]), z=array([6.514814, 6.613315]), radius=0.057275, length=0.098952, branch_order=0.0, branch_id=0.0, volume=0.00102, parent_id=43.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.009426000000000018, dy=-0.0005190000000001582, dz=0.09850099999999973, surface_area=0.035609800675355865, sa_to_vol=34.91156928956457, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=45.0, x=array([-0.17596, -0.15214]), y=array([-2.948445, -2.909786]), z=array([6.613315, 6.697043]), radius=0.043887, length=0.095249, branch_order=0.0, branch_id=0.0, volume=0.000576, parent_id=44.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.023820000000000008, dy=0.038659, dz=0.0837279999999998, surface_area=0.026264926377978573, sa_to_vol=45.59883051732391, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=46.0, x=array([-0.15214 , -0.137455]), y=array([-2.909786, -2.893629]), z=array([6.697043, 6.768243]), radius=0.051571, length=0.074472, branch_order=0.0, branch_id=0.0, volume=0.000622, parent_id=45.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.014685000000000004, dy=0.0161570000000002, dz=0.07120000000000015, surface_area=0.024131173291818256, sa_to_vol=38.79609853990073, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=47.0, x=array([-0.137455, -0.121855]), y=array([-2.893629, -2.87306 ]), z=array([6.768243, 6.835912]), radius=0.044826, length=0.072427, branch_order=0.0, branch_id=0.0, volume=0.000457, parent_id=46.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.015599999999999989, dy=0.020568999999999615, dz=0.06766900000000042, surface_area=0.020399069227309014, sa_to_vol=44.63691297004161, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=48.0, x=array([-0.121855, -0.105806]), y=array([-2.87306 , -2.850741]), z=array([6.835912, 6.914823]), radius=0.045028, length=0.083562, branch_order=0.0, branch_id=0.0, volume=0.000532, parent_id=47.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.016049000000000008, dy=0.022318999999999978, dz=0.07891099999999973, surface_area=0.0236412998735922, sa_to_vol=44.43853359697782, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=49.0, x=array([-0.105806, -0.106141]), y=array([-2.850741, -2.826305]), z=array([6.914823, 6.981695]), radius=0.044382, length=0.071198, branch_order=0.0, branch_id=0.0, volume=0.000441, parent_id=48.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.00033500000000000196, dy=0.024436000000000124, dz=0.06687200000000004, surface_area=0.019854297796930386, sa_to_vol=45.02108343975144, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=50.0, x=array([-0.106141, -0.107509]), y=array([-2.826305, -2.805509]), z=array([6.981695, 7.054587]), radius=0.051491, length=0.075813, branch_order=0.0, branch_id=0.0, volume=0.000631, parent_id=49.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0013679999999999942, dy=0.02079600000000026, dz=0.07289199999999951, surface_area=0.024527589952050875, sa_to_vol=38.8709824913643, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=51.0, x=array([-0.107509, -0.110174]), y=array([-2.805509, -2.787535]), z=array([7.054587, 7.135171]), radius=0.051472, length=0.082607, branch_order=0.0, branch_id=0.0, volume=0.000688, parent_id=50.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0026650000000000007, dy=0.017973999999999712, dz=0.08058399999999999, surface_area=0.026715774084031713, sa_to_vol=38.831066982604234, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=52.0, x=array([-0.110174, -0.102546]), y=array([-2.787535, -2.768837]), z=array([7.135171, 7.207126]), radius=0.046712, length=0.074735, branch_order=0.0, branch_id=0.0, volume=0.000512, parent_id=51.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.007627999999999996, dy=0.018698000000000103, dz=0.07195499999999999, surface_area=0.021934733864874684, sa_to_vol=42.84127707983337, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=53.0, x=array([-0.102546, -0.098577]), y=array([-2.768837, -2.739703]), z=array([7.207126, 7.284497]), radius=0.051431, length=0.08277, branch_order=0.0, branch_id=0.0, volume=0.000688, parent_id=52.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.003969, dy=0.029133999999999993, dz=0.0773710000000003, surface_area=0.026747167177472202, sa_to_vol=38.876696478884014, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=54.0, x=array([-0.098577, -0.081751]), y=array([-2.739703, -2.720897]), z=array([7.284497, 7.355692]), radius=0.044846, length=0.075535, branch_order=0.0, branch_id=0.0, volume=0.000477, parent_id=53.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.016825999999999994, dy=0.0188060000000001, dz=0.07119500000000034, surface_area=0.021283929636066068, sa_to_vol=44.62039755988693, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=55.0, x=array([-0.081751, -0.083484]), y=array([-2.720897, -2.701513]), z=array([7.355692, 7.432121]), radius=0.051389, length=0.078868, branch_order=0.0, branch_id=0.0, volume=0.000654, parent_id=54.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0017329999999999984, dy=0.019384000000000068, dz=0.07642900000000008, surface_area=0.025465421137814397, sa_to_vol=38.93795281011376, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=56.0, x=array([-0.083484, -0.07584 ]), y=array([-2.701513, -2.686353]), z=array([7.432121, 7.497618]), radius=0.051372, length=0.067662, branch_order=0.0, branch_id=0.0, volume=0.000561, parent_id=55.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.007643999999999998, dy=0.01515999999999984, dz=0.0654969999999997, surface_area=0.021839926529916273, sa_to_vol=38.930350320706374, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=57.0, x=array([-0.07584 , -0.059804]), y=array([-2.686353, -2.671483]), z=array([7.497618, 7.557284]), radius=0.05215, length=0.063547, branch_order=0.0, branch_id=0.0, volume=0.000543, parent_id=56.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.016036, dy=0.01487000000000016, dz=0.059666, surface_area=0.020822325625705046, sa_to_vol=38.34682435673121, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=58.0, x=array([-0.059804, -0.04742 ]), y=array([-2.671483, -2.643733]), z=array([7.557284, 7.617007]), radius=0.051333, length=0.06701, branch_order=0.0, branch_id=0.0, volume=0.000555, parent_id=57.0, reverse_branch_order=14.0, segment_id=9.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.012384000000000006, dy=0.02774999999999972, dz=0.05972299999999997, surface_area=0.021613053689534867, sa_to_vol=38.942439080243, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=59.0, x=array([-0.04742 , -0.056788]), y=array([-2.643733, -2.617914]), z=array([7.617007, 7.723978]), radius=0.047613, length=0.110441, branch_order=0.0, branch_id=0.0, volume=0.000787, parent_id=58.0, reverse_branch_order=13.0, segment_id=10.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.009368000000000001, dy=0.02581900000000026, dz=0.1069709999999997, surface_area=0.03303967335757714, sa_to_vol=41.98179588002178, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=60.0, x=array([-0.056788, -0.090898]), y=array([-2.617914, -2.551916]), z=array([7.723978, 7.775558]), radius=0.040756, length=0.090442, branch_order=0.0, branch_id=0.0, volume=0.000472, parent_id=59.0, reverse_branch_order=11.0, segment_id=11.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.03411000000000001, dy=0.065998, dz=0.0515800000000004, surface_area=0.023160161289314703, sa_to_vol=49.06813832481929, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=61.0, x=array([-0.090898, -0.106074]), y=array([-2.551916, -2.50664 ]), z=array([7.775558, 7.830104]), radius=0.038178, length=0.072495, branch_order=0.0, branch_id=0.0, volume=0.000332, parent_id=60.0, reverse_branch_order=11.0, segment_id=11.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.015175999999999995, dy=0.04527599999999987, dz=0.054546000000000205, surface_area=0.017390060630425624, sa_to_vol=52.379700694053085, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=62.0, x=array([-0.106074, -0.122123]), y=array([-2.50664 , -2.471698]), z=array([7.830104, 7.885584]), radius=0.038446, length=0.067503, branch_order=0.0, branch_id=0.0, volume=0.000313, parent_id=61.0, reverse_branch_order=11.0, segment_id=11.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.016048999999999994, dy=0.03494200000000003, dz=0.05547999999999931, surface_area=0.016306250296615234, sa_to_vol=52.09664631506464, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=63.0, x=array([-0.122123, -0.14285 ]), y=array([-2.471698, -2.427033]), z=array([7.885584, 7.934678]), radius=0.037605, length=0.069532, branch_order=0.0, branch_id=0.0, volume=0.000309, parent_id=62.0, reverse_branch_order=11.0, segment_id=11.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.02072700000000001, dy=0.04466499999999973, dz=0.04909400000000019, surface_area=0.016428964185487185, sa_to_vol=53.1681688850718, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=64.0, x=array([-0.14285 , -0.157797]), y=array([-2.427033, -2.393002]), z=array([7.934678, 7.993905]), radius=0.038695, length=0.069924, branch_order=0.0, branch_id=0.0, volume=0.000329, parent_id=63.0, reverse_branch_order=11.0, segment_id=11.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.014946999999999988, dy=0.034031000000000144, dz=0.05922699999999992, surface_area=0.017000472165276927, sa_to_vol=51.67316767561376, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=65.0, x=array([-0.157797, -0.168682]), y=array([-2.393002, -2.381461]), z=array([7.993905, 8.042353]), radius=0.038695, length=0.050979, branch_order=0.0, branch_id=0.0, volume=0.00024, parent_id=64.0, reverse_branch_order=11.0, segment_id=11.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.010885000000000006, dy=0.011541000000000246, dz=0.04844800000000049, surface_area=0.012394414943562332, sa_to_vol=51.64339559817638, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=66.0, x=array([-0.168682, -0.198118]), y=array([-2.381461, -2.316807]), z=array([8.042353, 8.10925 ]), radius=0.036929, length=0.097579, branch_order=0.0, branch_id=0.0, volume=0.000418, parent_id=65.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.02943599999999999, dy=0.06465399999999999, dz=0.0668969999999991, surface_area=0.0226414261536279, sa_to_vol=54.16609127662177, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=67.0, x=array([-0.198118, -0.216827]), y=array([-2.316807, -2.290872]), z=array([8.10925 , 8.168945]), radius=0.035127, length=0.067721, branch_order=0.0, branch_id=0.0, volume=0.000263, parent_id=66.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.018709000000000003, dy=0.02593500000000004, dz=0.059695000000001386, surface_area=0.01494666468277062, sa_to_vol=56.831424649317945, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=68.0, x=array([-0.216827, -0.234318]), y=array([-2.290872, -2.259834]), z=array([8.168945, 8.232016]), radius=0.036007, length=0.072439, branch_order=0.0, branch_id=0.0, volume=0.000295, parent_id=67.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.017491000000000007, dy=0.031037999999999677, dz=0.06307099999999899, surface_area=0.01638850181042742, sa_to_vol=55.55424342517769, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=69.0, x=array([-0.234318, -0.249533]), y=array([-2.259834, -2.230372]), z=array([8.232016, 8.288784]), radius=0.036566, length=0.065742, branch_order=0.0, branch_id=0.0, volume=0.000276, parent_id=68.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.015215000000000006, dy=0.0294620000000001, dz=0.05676799999999993, surface_area=0.015104287214076575, sa_to_vol=54.725678311871654, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=70.0, x=array([-0.249533, -0.268207]), y=array([-2.230372, -2.2044  ]), z=array([8.288784, 8.352136]), radius=0.036996, length=0.070971, branch_order=0.0, branch_id=0.0, volume=0.000305, parent_id=69.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.01867399999999997, dy=0.025971999999999884, dz=0.06335200000000007, surface_area=0.016497402248348426, sa_to_vol=54.08984343720796, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=71.0, x=array([-0.268207, -0.286132]), y=array([-2.2044  , -2.176682]), z=array([8.352136, 8.403738]), radius=0.035783, length=0.061256, branch_order=0.0, branch_id=0.0, volume=0.000246, parent_id=70.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.017925000000000024, dy=0.02771800000000013, dz=0.051602000000000814, surface_area=0.01377226120293602, sa_to_vol=55.98480163795129, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=72.0, x=array([-0.286132, -0.303921]), y=array([-2.176682, -2.142636]), z=array([8.403738, 8.473344]), radius=0.036562, length=0.079502, branch_order=0.0, branch_id=0.0, volume=0.000334, parent_id=71.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.017789, dy=0.03404600000000002, dz=0.06960600000000028, surface_area=0.018263662237129854, sa_to_vol=54.68162346446064, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=73.0, x=array([-0.303921, -0.315822]), y=array([-2.142636, -2.117104]), z=array([8.473344, 8.539532]), radius=0.042866, length=0.071933, branch_order=0.0, branch_id=0.0, volume=0.000415, parent_id=72.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.011900999999999995, dy=0.02553200000000011, dz=0.06618799999999858, surface_area=0.019374076092752032, sa_to_vol=46.684520705426586, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=74.0, x=array([-0.315822, -0.335579]), y=array([-2.117104, -2.071696]), z=array([8.539532, 8.600119]), radius=0.037867, length=0.07825, branch_order=0.0, branch_id=0.0, volume=0.000352, parent_id=73.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.019757000000000025, dy=0.04540799999999967, dz=0.060586999999999946, surface_area=0.018617660830610352, sa_to_vol=52.89108190514305, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=75.0, x=array([-0.335579, -0.347286]), y=array([-2.071696, -2.045686]), z=array([8.600119, 8.666266]), radius=0.039886, length=0.072034, branch_order=0.0, branch_id=0.0, volume=0.00036, parent_id=74.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.011706999999999967, dy=0.02601000000000031, dz=0.06614700000000084, surface_area=0.018052522078067387, sa_to_vol=50.14589466129829, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=76.0, x=array([-0.347286, -0.363641]), y=array([-2.045686, -2.024581]), z=array([8.666266, 8.731817]), radius=0.044711, length=0.07078, branch_order=0.0, branch_id=0.0, volume=0.000445, parent_id=75.0, reverse_branch_order=10.0, segment_id=12.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.01635500000000001, dy=0.02110499999999993, dz=0.06555099999999925, surface_area=0.019884048327501513, sa_to_vol=44.68325466854273, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=77.0, x=array([-0.363641, -0.367266]), y=array([-2.024581, -1.956941]), z=array([8.731817, 8.793305]), radius=0.03301, length=0.091483, branch_order=0.0, branch_id=0.0, volume=0.000313, parent_id=76.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0036249999999999893, dy=0.06763999999999992, dz=0.06148800000000065, surface_area=0.018974301214485997, sa_to_vol=60.62077065330989, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=78.0, x=array([-0.367266, -0.368869]), y=array([-1.956941, -1.924685]), z=array([8.793305, 8.837513]), radius=0.032015, length=0.054749, branch_order=0.0, branch_id=0.0, volume=0.000176, parent_id=77.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.001603000000000021, dy=0.03225600000000006, dz=0.04420799999999936, surface_area=0.011013099567934549, sa_to_vol=62.57442936326448, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=79.0, x=array([-0.368869, -0.374497]), y=array([-1.924685, -1.884855]), z=array([8.837513, 8.885968]), radius=0.03024, length=0.062976, branch_order=0.0, branch_id=0.0, volume=0.000181, parent_id=78.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.005628000000000022, dy=0.03983000000000003, dz=0.04845500000000058, surface_area=0.011965661907845435, sa_to_vol=66.1086293251129, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=80.0, x=array([-0.374497, -0.376017]), y=array([-1.884855, -1.84986 ]), z=array([8.885968, 8.937917]), radius=0.028642, length=0.062655, branch_order=0.0, branch_id=0.0, volume=0.000161, parent_id=79.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0015199999999999658, dy=0.03499499999999989, dz=0.05194900000000047, surface_area=0.011275581362017936, sa_to_vol=70.0346668448319, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=81.0, x=array([-0.376017, -0.382543]), y=array([-1.84986 , -1.822089]), z=array([8.937917, 8.988857]), radius=0.034786, length=0.058383, branch_order=0.0, branch_id=0.0, volume=0.000222, parent_id=80.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.006526000000000032, dy=0.02777099999999999, dz=0.050939999999998875, surface_area=0.01276059039415044, sa_to_vol=57.48013691058757, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=82.0, x=array([-0.382543, -0.38097 ]), y=array([-1.822089, -1.788348]), z=array([8.988857, 9.050148]), radius=0.032496, length=0.069983, branch_order=0.0, branch_id=0.0, volume=0.000232, parent_id=81.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0015730000000000466, dy=0.03374100000000002, dz=0.06129100000000065, surface_area=0.014289016249321931, sa_to_vol=61.59058728156005, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=83.0, x=array([-0.38097 , -0.404069]), y=array([-1.788348, -1.738043]), z=array([9.050148, 9.096119]), radius=0.030276, length=0.071954, branch_order=0.0, branch_id=0.0, volume=0.000207, parent_id=82.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.023099000000000036, dy=0.050305000000000044, dz=0.04597099999999976, surface_area=0.013687789154887612, sa_to_vol=66.12458528931214, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=84.0, x=array([-0.404069, -0.419862]), y=array([-1.738043, -1.720933]), z=array([9.096119, 9.157446]), radius=0.029328, length=0.065598, branch_order=0.0, branch_id=0.0, volume=0.000177, parent_id=83.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.015793, dy=0.01710999999999996, dz=0.06132700000000035, surface_area=0.012087957223478593, sa_to_vol=68.29354363547228, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=85.0, x=array([-0.419862, -0.42789 ]), y=array([-1.720933, -1.685086]), z=array([9.157446, 9.214547]), radius=0.033145, length=0.067897, branch_order=0.0, branch_id=0.0, volume=0.000234, parent_id=84.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.00802799999999998, dy=0.03584699999999996, dz=0.0571009999999994, surface_area=0.014139969650208115, sa_to_vol=60.42722072738511, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=86.0, x=array([-0.42789 , -0.446164]), y=array([-1.685086, -1.649179]), z=array([9.214547, 9.25704 ]), radius=0.036584, length=0.058557, branch_order=0.0, branch_id=0.0, volume=0.000246, parent_id=85.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.018274000000000012, dy=0.03590700000000013, dz=0.042493000000000336, surface_area=0.01346014925067753, sa_to_vol=54.7160538645428, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=87.0, x=array([-0.446164, -0.462466]), y=array([-1.649179, -1.588237]), z=array([9.25704 , 9.302596]), radius=0.031391, length=0.077814, branch_order=0.0, branch_id=0.0, volume=0.000241, parent_id=86.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.016301999999999983, dy=0.06094200000000005, dz=0.045555999999999486, surface_area=0.015347680860842754, sa_to_vol=63.68332307403632, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=88.0, x=array([-0.462466, -0.4719  ]), y=array([-1.588237, -1.552052]), z=array([9.302596, 9.340372]), radius=0.037065, length=0.053154, branch_order=0.0, branch_id=0.0, volume=0.000229, parent_id=87.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.009433999999999998, dy=0.03618499999999991, dz=0.03777600000000092, surface_area=0.012378836445327635, sa_to_vol=54.056054346408885, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=89.0, x=array([-0.4719  , -0.511085]), y=array([-1.552052, -1.491902]), z=array([9.340372, 9.389201]), radius=0.032378, length=0.08682, branch_order=0.0, branch_id=0.0, volume=0.000286, parent_id=88.0, reverse_branch_order=9.0, segment_id=13.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.039185000000000025, dy=0.060149999999999926, dz=0.04882899999999957, surface_area=0.017662398071902223, sa_to_vol=61.75663661504274, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=90.0, x=array([-0.511085, -0.508096]), y=array([-1.491902, -1.473524]), z=array([9.389201, 9.448287]), radius=0.032622, length=0.06195, branch_order=0.0, branch_id=0.0, volume=0.000207, parent_id=89.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0029890000000000194, dy=0.018378000000000005, dz=0.05908600000000064, surface_area=0.012697895904075829, sa_to_vol=61.3424922902214, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=91.0, x=array([-0.508096, -0.489413]), y=array([-1.473524, -1.45054 ]), z=array([9.448287, 9.508954]), radius=0.028499, length=0.067511, branch_order=0.0, branch_id=0.0, volume=0.000172, parent_id=90.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.018683000000000005, dy=0.022984000000000115, dz=0.060666999999998694, surface_area=0.012088823329157256, sa_to_vol=70.28385656486776, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=92.0, x=array([-0.489413, -0.496181]), y=array([-1.45054 , -1.422676]), z=array([9.508954, 9.569169]), radius=0.027423, length=0.066694, branch_order=0.0, branch_id=0.0, volume=0.000158, parent_id=91.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.006767999999999996, dy=0.02786399999999989, dz=0.06021500000000124, surface_area=0.011491629015530939, sa_to_vol=72.73182921222114, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=93.0, x=array([-0.496181, -0.499306]), y=array([-1.422676, -1.402556]), z=array([9.569169, 9.61718 ]), radius=0.032231, length=0.05215, branch_order=0.0, branch_id=0.0, volume=0.00017, parent_id=92.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0031250000000000444, dy=0.020120000000000138, dz=0.048010999999998916, surface_area=0.010561070974902033, sa_to_vol=62.123946911188426, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=94.0, x=array([-0.499306, -0.501735]), y=array([-1.402556, -1.383654]), z=array([9.61718 , 9.681935]), radius=0.029584, length=0.067501, branch_order=0.0, branch_id=0.0, volume=0.000186, parent_id=93.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.0024290000000000145, dy=0.018901999999999974, dz=0.0647549999999999, surface_area=0.01254720428536719, sa_to_vol=67.45808755573758, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=95.0, x=array([-0.501735, -0.485299]), y=array([-1.383654, -1.349411]), z=array([9.681935, 9.737485]), radius=0.032202, length=0.067294, branch_order=0.0, branch_id=0.0, volume=0.000219, parent_id=94.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.01643600000000006, dy=0.03424300000000002, dz=0.05555000000000021, surface_area=0.01361567128171937, sa_to_vol=62.17201498501995, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=96.0, x=array([-0.485299, -0.473887]), y=array([-1.349411, -1.292763]), z=array([9.737485, 9.808151]), radius=0.032156, length=0.091286, branch_order=0.0, branch_id=0.0, volume=0.000297, parent_id=95.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.011411999999999978, dy=0.05664799999999981, dz=0.070666000000001, surface_area=0.018443615755654647, sa_to_vol=62.09971634900554, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=97.0, x=array([-0.473887, -0.509462]), y=array([-1.292763, -1.247464]), z=array([9.808151, 9.88034 ]), radius=0.031997, length=0.092352, branch_order=0.0, branch_id=0.0, volume=0.000297, parent_id=96.0, reverse_branch_order=8.0, segment_id=14.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.03557499999999997, dy=0.0452990000000002, dz=0.07218899999999984, surface_area=0.018566730549448306, sa_to_vol=62.51424427423672, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=98.0, x=array([-0.509462, -0.46907 ]), y=array([-1.247464, -1.110556]), z=array([9.88034 , 9.935555]), radius=0.026386, length=0.153049, branch_order=0.0, branch_id=0.0, volume=0.000335, parent_id=97.0, reverse_branch_order=7.0, segment_id=15.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.040391999999999983, dy=0.1369079999999998, dz=0.05521500000000046, surface_area=0.02537370712808005, sa_to_vol=75.74240933755239, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=99.0, x=array([-0.46907 , -0.494318]), y=array([-1.110556, -1.038573]), z=array([ 9.935555, 10.002307]), radius=0.026233, length=0.101364, branch_order=0.0, branch_id=0.0, volume=0.000219, parent_id=98.0, reverse_branch_order=7.0, segment_id=15.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.025247999999999993, dy=0.07198300000000013, dz=0.06675199999999926, surface_area=0.01670750377174687, sa_to_vol=76.28997156048798, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=100.0, x=array([-0.494318, -0.501549]), y=array([-1.038573, -1.036312]), z=array([10.002307, 10.092412]), radius=0.024533, length=0.090423, branch_order=0.0, branch_id=0.0, volume=0.000171, parent_id=99.0, reverse_branch_order=6.0, segment_id=16.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.007231000000000043, dy=0.0022610000000000685, dz=0.09010499999999944, surface_area=0.013938288160607971, sa_to_vol=81.51045707957877, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=101.0, x=array([-0.501549, -0.440411]), y=array([-1.036312, -1.053649]), z=array([10.092412, 10.163284]), radius=0.019178, length=0.095191, branch_order=0.0, branch_id=0.0, volume=0.00011, parent_id=100.0, reverse_branch_order=5.0, segment_id=17.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.061138000000000026, dy=-0.017337000000000158, dz=0.07087200000000138, surface_area=0.011470413438217388, sa_to_vol=104.27648580197625, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=102.0, x=array([-0.440411, -0.437682]), y=array([-1.053649, -1.029842]), z=array([10.163284, 10.217745]), radius=0.019129, length=0.0595, branch_order=0.0, branch_id=0.0, volume=6.8e-05, parent_id=101.0, reverse_branch_order=5.0, segment_id=17.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0027289999999999814, dy=0.023807000000000134, dz=0.05446099999999987, surface_area=0.00715136757859178, sa_to_vol=105.16717027340853, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=103.0, x=array([-0.437682, -0.423197]), y=array([-1.029842, -1.00204 ]), z=array([10.217745, 10.27936 ]), radius=0.019104, length=0.069132, branch_order=0.0, branch_id=0.0, volume=7.9e-05, parent_id=102.0, reverse_branch_order=5.0, segment_id=17.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.014485000000000026, dy=0.027801999999999882, dz=0.06161499999999975, surface_area=0.00829818855979506, sa_to_vol=105.04036151639318, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=104.0, x=array([-0.423197, -0.383383]), y=array([-1.00204 , -0.969435]), z=array([10.27936 , 10.339124]), radius=0.01905, length=0.078867, branch_order=0.0, branch_id=0.0, volume=9e-05, parent_id=103.0, reverse_branch_order=5.0, segment_id=17.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.039814000000000016, dy=0.032604999999999995, dz=0.059763999999999484, surface_area=0.009439960335586382, sa_to_vol=104.88844817318201, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=105.0, x=array([-0.383383, -0.363553]), y=array([-0.969435, -1.01903 ]), z=array([10.339124, 10.495052]), radius=0.018641, length=0.164822, branch_order=0.0, branch_id=0.0, volume=0.00018, parent_id=104.0, reverse_branch_order=4.0, segment_id=18.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.01982999999999996, dy=-0.049595000000000056, dz=0.1559279999999994, surface_area=0.01930475323173584, sa_to_vol=107.2486290651991, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=106.0, x=array([-0.363553, -0.332222]), y=array([-1.01903 , -1.151523]), z=array([10.495052, 10.622299]), radius=0.017021, length=0.186353, branch_order=0.0, branch_id=0.0, volume=0.00017, parent_id=105.0, reverse_branch_order=3.0, segment_id=19.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.031331, dy=-0.13249299999999997, dz=0.12724700000000055, surface_area=0.019929726035392767, sa_to_vol=117.23368256113392, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=138.0, x=array([-0.383383, -0.376465]), y=array([-0.969435, -0.878738]), z=array([10.339124, 10.340219]), radius=0.004507, length=0.090968, branch_order=1.0, branch_id=4.0, volume=6e-06, parent_id=104.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.00691799999999998, dy=0.09069700000000003, dz=0.001094999999999402, surface_area=0.0025760605862129708, sa_to_vol=429.34343103549514, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=139.0, x=array([-0.376465, -0.421839]), y=array([-0.878738, -0.783571]), z=array([10.340219, 10.350996]), radius=0.004334, length=0.105979, branch_order=1.0, branch_id=4.0, volume=6e-06, parent_id=138.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=-0.045374000000000025, dy=0.095167, dz=0.010777000000000925, surface_area=0.0028859486050319834, sa_to_vol=480.99143417199724, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=140.0, x=array([-0.421839, -0.342949]), y=array([-0.783571, -0.749089]), z=array([10.350996, 10.412333]), radius=0.004115, length=0.105711, branch_order=1.0, branch_id=4.0, volume=6e-06, parent_id=139.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.07889000000000002, dy=0.03448200000000001, dz=0.061336999999999975, surface_area=0.0027331904152598794, sa_to_vol=455.53173587664656, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=141.0, x=array([-0.342949, -0.279745]), y=array([-0.749089, -0.658196]), z=array([10.412333, 10.448264]), radius=0.003875, length=0.116393, branch_order=1.0, branch_id=4.0, volume=5e-06, parent_id=140.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.06320399999999998, dy=0.090893, dz=0.03593099999999971, surface_area=0.002833860301401895, sa_to_vol=566.7720602803789, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=142.0, x=array([-0.279745, -0.246297]), y=array([-0.658196, -0.599935]), z=array([10.448264, 10.49903 ]), radius=0.003577, length=0.084204, branch_order=1.0, branch_id=4.0, volume=3e-06, parent_id=141.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.03344800000000003, dy=0.05826100000000001, dz=0.05076599999999942, surface_area=0.0018924810134617672, sa_to_vol=630.8270044872557, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=143.0, x=array([-0.246297, -0.199178]), y=array([-0.599935, -0.539375]), z=array([10.49903 , 10.507596]), radius=0.00333, length=0.077208, branch_order=1.0, branch_id=4.0, volume=3e-06, parent_id=142.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.047118999999999994, dy=0.06055999999999995, dz=0.008566000000000074, surface_area=0.0016154235300850825, sa_to_vol=538.4745100283608, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=144.0, x=array([-0.199178, -0.123813]), y=array([-0.539375, -0.497206]), z=array([10.507596, 10.572537]), radius=0.002566, length=0.108053, branch_order=1.0, branch_id=4.0, volume=2e-06, parent_id=143.0, reverse_branch_order=1.0, segment_id=29.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.07536499999999999, dy=0.04216900000000007, dz=0.06494100000000103, surface_area=0.00174210107844347, sa_to_vol=871.0505392217351, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=145.0, x=array([-0.123813, -0.051498]), y=array([-0.497206, -0.454615]), z=array([10.572537, 10.598556]), radius=0.0025, length=0.087866, branch_order=1.0, branch_id=4.0, volume=2e-06, parent_id=144.0, reverse_branch_order=1.0, segment_id=29.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.072315, dy=0.04259099999999999, dz=0.026018999999999792, surface_area=0.001380195900501604, sa_to_vol=690.097950250802, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=146.0, x=array([-0.051498,  0.052574]), y=array([-0.454615, -0.451659]), z=array([10.598556, 10.618096]), radius=0.0025, length=0.105932, branch_order=1.0, branch_id=4.0, volume=2e-06, parent_id=145.0, reverse_branch_order=1.0, segment_id=29.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.104072, dy=0.002956000000000014, dz=0.019539999999999225, surface_area=0.00166397596490037, sa_to_vol=831.987982450185, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=147.0, x=array([0.052574, 0.07618 ]), y=array([-0.451659, -0.383481]), z=array([10.618096, 10.653787]), radius=0.0025, length=0.080494, branch_order=1.0, branch_id=4.0, volume=2e-06, parent_id=146.0, reverse_branch_order=1.0, segment_id=29.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.023605999999999995, dy=0.06817799999999996, dz=0.03569099999999992, surface_area=0.0012643967952902842, sa_to_vol=632.1983976451421, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=148.0, x=array([-0.437682, -0.387682]), y=array([-1.029842, -0.979842]), z=array([10.217745, 10.117745]), radius=0.004507, length=0.090968, branch_order=1.0, branch_id=4.0, volume=6e-06, parent_id=102.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.04999999999999999, dy=0.04999999999999993, dz=-0.10000000000000142, surface_area=0.0025760605862129708, sa_to_vol=429.34343103549514, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=149.0, x=array([-0.387682, -0.337682]), y=array([-0.979842, -0.929842]), z=array([10.117745, 10.017745]), radius=0.004334, length=0.105979, branch_order=1.0, branch_id=4.0, volume=6e-06, parent_id=148.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.050000000000000044, dy=0.050000000000000044, dz=-0.09999999999999964, surface_area=0.0028859486050319834, sa_to_vol=480.99143417199724, slope=0.0, is_stem=False, is_divide=False)",
-    "Cylinder(cyl_id=150.0, x=array([-0.337682, -0.287682]), y=array([-0.929842, -0.879842]), z=array([10.017745,  9.917745]), radius=0.004115, length=0.105711, branch_order=1.0, branch_id=4.0, volume=6e-06, parent_id=149.0, reverse_branch_order=2.0, segment_id=28.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.04999999999999999, dy=0.04999999999999993, dz=-0.09999999999999964, surface_area=0.0027331904152598794, sa_to_vol=455.53173587664656, slope=0.0, is_stem=False, is_divide=False)",
-]
-
-ez_projection_xy = [0.7996, -0.7996, -0.7996, -0.7996, -0.7996]
-ez_projection_xz = [0.6405, 0.6405, 0.6405, 0.6405, -0.6405]
-ez_projection_yz = [0.3667, 0.3667, 0.3667, -0.3667, 0.3667]
-
-ten_cyls_id_one = [1.0]
-ten_cyls_bo_one = [7.0, 8.0, 9.0]
-ten_cyls_bo_and_len = [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 8.0]
-
-ten_cyls_is_stem = (True, True, True, True, True, True, True, True, True, True)
-happy_path_is_stem = (
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-)
-
-small_tree_is_stem = (
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    True,
-    False,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-)
-
-drip_mid_stem_map = (
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-)
-
-drip_adj_stem_map = (
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-)
-
-drip_on_trunk_stem_map = (
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    True,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-    False,
-)
-
-ten_cyls_rows_proj_XZ = []
-ten_cyls_rows_proj_XY = []
-ten_cyls_rows_proj_YZ = []
-
-ten_cyls_flows = [
-    Flow(
-        num_cylinders=10,
-        projected_area=1.332563385222201,
-        surface_area=3.3102009951007805,
-        angle_sum=13.311017528987755,
-        volume=0.327783,
-        sa_to_vol=100.98563451833446,
-        drip_node_id=0,
-        drip_node_loc=(0.076822, -3.386935),
-    )
-]
-
-happy_path_flows = [
-    Flow(
-        num_cylinders=117,
-        projected_area=2.966550840828153,
-        surface_area=7.4149259812596755,
-        angle_sum=127.43768060688681,
-        volume=0.5217939999999999,
-        sa_to_vol=10791.467322627153,
-        drip_node_id=0,
-        drip_node_loc=(0.076822, -3.386935),
-    ),
-    Flow(
-        num_cylinders=3,
-        projected_area=0.001975969393493081,
-        surface_area=0.008195199606504833,
-        angle_sum=-2.8659498543735316,
-        volume=1.8e-05,
-        sa_to_vol=1365.866601084139,
-        drip_node_id=149.0,
-        drip_node_loc=(-0.337682, -0.929842, 10.017745),
-    ),
-]
-
-small_tree_flows = [
-    Flow(
-        num_cylinders=216,
-        projected_area=16.517060215326925,
-        surface_area=19.49507498855058,
-        angle_sum=180.03288665020412,
-        volume=3.9062959999999998,
-        sa_to_vol=83646.4965224844,
-        drip_node_id=0,
-        drip_node_loc=(-0.299115, 2.537844),
-    ),
-    Flow(
-        num_cylinders=59,
-        projected_area=0.01712513891199792,
-        surface_area=0.0653136170203522,
-        angle_sum=4.0063304619261375,
-        volume=7.699999999999999e-05,
-        sa_to_vol=52984.87647431056,
-        drip_node_id=405.0,
-        drip_node_loc=(1.931817, 2.350368, 13.193707),
-    ),
-    Flow(
-        num_cylinders=32,
-        projected_area=0.01033455063954463,
-        surface_area=0.03666120817014779,
-        angle_sum=3.2693589141152817,
-        volume=4.8e-05,
-        sa_to_vol=25387.06100800808,
-        drip_node_id=222.0,
-        drip_node_loc=(1.607588, 2.540239, 13.949144),
-    ),
-    Flow(
-        num_cylinders=4,
-        projected_area=0.001879242349756744,
-        surface_area=0.006360028663486393,
-        angle_sum=-1.4004883938106008,
-        volume=7e-06,
-        sa_to_vol=2637.736169825428,
-        drip_node_id=188.0,
-        drip_node_loc=(1.996825, 2.461936, 14.202651),
-    ),
-    Flow(
-        num_cylinders=8,
-        projected_area=0.0024559778348642487,
-        surface_area=0.008355741104642577,
-        angle_sum=-2.5182426363497505,
-        volume=9.999999999999999e-06,
-        sa_to_vol=6887.50207002412,
-        drip_node_id=147.0,
-        drip_node_loc=(1.704854, 2.706883, 14.222385),
-    ),
-    Flow(
-        num_cylinders=6,
-        projected_area=0.0016837038216656593,
-        surface_area=0.005825910788486325,
-        angle_sum=-2.115709206226209,
-        volume=5.999999999999999e-06,
-        sa_to_vol=5825.910788486325,
-        drip_node_id=163.0,
-        drip_node_loc=(1.097889, 2.391842, 14.289977),
-    ),
-    Flow(
-        num_cylinders=3,
-        projected_area=0.0009449634116903826,
-        surface_area=0.003272345739832201,
-        angle_sum=0.3639153227121651,
-        volume=4e-06,
-        sa_to_vol=2617.2529857240233,
-        drip_node_id=176.0,
-        drip_node_loc=(1.730164, 2.302889, 14.604732),
-    ),
-    Flow(
-        num_cylinders=5,
-        projected_area=0.00142245008970565,
-        surface_area=0.00503121351083425,
-        angle_sum=-0.00046949725728318015,
-        volume=7e-06,
-        sa_to_vol=3582.900027621186,
-        drip_node_id=179.0,
-        drip_node_loc=(1.652252, 2.280224, 14.47647),
-    ),
-    Flow(
-        num_cylinders=1,
-        projected_area=0.000277103153591026,
-        surface_area=0.0009704536786571552,
-        angle_sum=-0.525305433679695,
-        volume=1e-06,
-        sa_to_vol=970.4536786571553,
-        drip_node_id=226.0,
-        drip_node_loc=(1.844739, 2.507194, 13.979754),
-    ),
-    Flow(
-        num_cylinders=35,
-        projected_area=0.01052004847925714,
-        surface_area=0.0436400206306485,
-        angle_sum=10.193206366817037,
-        volume=5.299999999999999e-05,
-        sa_to_vol=28178.673695003272,
-        drip_node_id=232.0,
-        drip_node_loc=(1.920238, 2.162247, 13.938462),
-    ),
-    Flow(
-        num_cylinders=3,
-        projected_area=0.0007556341840308715,
-        surface_area=0.0030501880153335982,
-        angle_sum=-1.0835316994580007,
-        volume=4e-06,
-        sa_to_vol=2407.4260123907607,
-        drip_node_id=274.0,
-        drip_node_loc=(2.065614, 2.478043, 14.019069),
-    ),
-    Flow(
-        num_cylinders=24,
-        projected_area=0.00687016322759011,
-        surface_area=0.02719439569967338,
-        angle_sum=-0.467953573105651,
-        volume=3.4e-05,
-        sa_to_vol=20303.90126264672,
-        drip_node_id=350.0,
-        drip_node_loc=(2.023443, 1.985047, 13.54066),
-    ),
-    Flow(
-        num_cylinders=5,
-        projected_area=0.0015186890252737617,
-        surface_area=0.005638373415030282,
-        angle_sum=-2.4195862206414995,
-        volume=7e-06,
-        sa_to_vol=4248.493555173988,
-        drip_node_id=309.0,
-        drip_node_loc=(2.154494, 1.637801, 13.701281),
-    ),
-    Flow(
-        num_cylinders=2,
-        projected_area=0.0006559427591282863,
-        surface_area=0.003129183362608114,
-        angle_sum=-1.7050548651799198,
-        volume=3e-06,
-        sa_to_vol=2136.118070826746,
-        drip_node_id=304.0,
-        drip_node_loc=(2.203542, 1.577092, 13.844742),
-    ),
-    Flow(
-        num_cylinders=5,
-        projected_area=0.0012687423706050432,
-        surface_area=0.006314459862046074,
-        angle_sum=-4.00094773716074,
-        volume=6.999999999999999e-06,
-        sa_to_vol=3509.096162206728,
-        drip_node_id=321.0,
-        drip_node_loc=(2.160463, 1.791407, 13.75528),
-    ),
-    Flow(
-        num_cylinders=3,
-        projected_area=0.0009350371357500096,
-        surface_area=0.0036045534549860535,
-        angle_sum=0.5863978722482486,
-        volume=4e-06,
-        sa_to_vol=2774.929521007695,
-        drip_node_id=315.0,
-        drip_node_loc=(2.164648, 1.893389, 13.78743),
-    ),
-    Flow(
-        num_cylinders=3,
-        projected_area=0.0008613860036361682,
-        surface_area=0.0030539265105913706,
-        angle_sum=-1.3293928778054518,
-        volume=4e-06,
-        sa_to_vol=2395.079553262153,
-        drip_node_id=325.0,
-        drip_node_loc=(1.842594, 1.682903, 13.750329),
-    ),
-    Flow(
-        num_cylinders=103,
-        projected_area=0.035842410317746466,
-        surface_area=0.1308078996954519,
-        angle_sum=0.9595299834971135,
-        volume=0.000184,
-        sa_to_vol=81927.57367017229,
-        drip_node_id=513.0,
-        drip_node_loc=(1.581945, 2.857679, 12.955942),
-    ),
-]
-
-drip_mid_flows = [
-    Flow(
-        num_cylinders=117,
-        projected_area=16.47655636448612,
-        surface_area=19.350321212472778,
-        angle_sum=134.88157931803016,
-        volume=3.906033,
-        sa_to_vol=7306.133204864248,
-        drip_node_id=0,
-        drip_node_loc=(-0.299115, 2.537844),
-    ),
-    Flow(
-        num_cylinders=14,
-        projected_area=0.010439067766281047,
-        surface_area=0.03744675454760344,
-        angle_sum=8.673607609925796,
-        volume=0.000121,
-        sa_to_vol=4804.1689124433915,
-        drip_node_id=117.0,
-        drip_node_loc=(1.011551, 2.485898, 13.266762),
-    ),
-]
-
-drip_adj_flows = [
-    Flow(
-        num_cylinders=107,
-        projected_area=2.9604039491372682,
-        surface_area=7.394338347069087,
-        angle_sum=124.04233409518383,
-        volume=0.521757,
-        sa_to_vol=4664.192277179154,
-        drip_node_id=0,
-        drip_node_loc=(0.076822, -3.386935),
-    ),
-    Flow(
-        num_cylinders=10,
-        projected_area=0.006539752252989748,
-        surface_area=0.027999615993246064,
-        angle_sum=7.133359698969585,
-        volume=5.2e-05,
-        sa_to_vol=6917.96517821172,
-        drip_node_id=125.0,
-        drip_node_loc=(-0.170586, -1.090298, 10.411159),
-    ),
-]
-
-ten_cyls_edges = [
-    (0.0, -1.0),
-    (0.0, 1.0),
-    (1.0, 2.0),
-    (2.0, 3.0),
-    (3.0, 4.0),
-    (4.0, 5.0),
-    (5.0, 6.0),
-    (6.0, 7.0),
-    (7.0, 8.0),
-    (8.0, 9.0),
-]
-
-happy_path_edges = [
-    (0.0, -1.0),
-    (0.0, 1.0),
-    (1.0, 2.0),
-    (2.0, 3.0),
-    (3.0, 4.0),
-    (4.0, 5.0),
-    (5.0, 6.0),
-    (6.0, 7.0),
-    (7.0, 8.0),
-    (8.0, 9.0),
-    (9.0, 10.0),
-    (10.0, 11.0),
-    (11.0, 12.0),
-    (12.0, 13.0),
-    (13.0, 14.0),
-    (14.0, 15.0),
-    (15.0, 16.0),
-    (16.0, 17.0),
-    (17.0, 18.0),
-    (18.0, 19.0),
-    (19.0, 20.0),
-    (20.0, 21.0),
-    (21.0, 22.0),
-    (22.0, 23.0),
-    (23.0, 24.0),
-    (24.0, 25.0),
-    (25.0, 26.0),
-    (26.0, 27.0),
-    (27.0, 28.0),
-    (28.0, 29.0),
-    (29.0, 30.0),
-    (30.0, 31.0),
-    (31.0, 32.0),
-    (32.0, 33.0),
-    (33.0, 34.0),
-    (34.0, 35.0),
-    (35.0, 36.0),
-    (36.0, 37.0),
-    (37.0, 38.0),
-    (38.0, 39.0),
-    (39.0, 40.0),
-    (40.0, 41.0),
-    (41.0, 42.0),
-    (42.0, 43.0),
-    (43.0, 44.0),
-    (44.0, 45.0),
-    (45.0, 46.0),
-    (46.0, 47.0),
-    (47.0, 48.0),
-    (48.0, 49.0),
-    (49.0, 50.0),
-    (50.0, 51.0),
-    (51.0, 52.0),
-    (52.0, 53.0),
-    (53.0, 54.0),
-    (54.0, 55.0),
-    (55.0, 56.0),
-    (56.0, 57.0),
-    (57.0, 58.0),
-    (58.0, 59.0),
-    (59.0, 60.0),
-    (60.0, 61.0),
-    (61.0, 62.0),
-    (62.0, 63.0),
-    (63.0, 64.0),
-    (64.0, 65.0),
-    (65.0, 66.0),
-    (66.0, 67.0),
-    (67.0, 68.0),
-    (68.0, 69.0),
-    (69.0, 70.0),
-    (70.0, 71.0),
-    (71.0, 72.0),
-    (72.0, 73.0),
-    (73.0, 74.0),
-    (74.0, 75.0),
-    (75.0, 76.0),
-    (76.0, 77.0),
-    (77.0, 78.0),
-    (78.0, 79.0),
-    (79.0, 80.0),
-    (80.0, 81.0),
-    (81.0, 82.0),
-    (82.0, 83.0),
-    (83.0, 84.0),
-    (84.0, 85.0),
-    (85.0, 86.0),
-    (86.0, 87.0),
-    (87.0, 88.0),
-    (88.0, 89.0),
-    (89.0, 90.0),
-    (90.0, 91.0),
-    (91.0, 92.0),
-    (92.0, 93.0),
-    (93.0, 94.0),
-    (94.0, 95.0),
-    (95.0, 96.0),
-    (96.0, 97.0),
-    (97.0, 98.0),
-    (98.0, 99.0),
-    (99.0, 100.0),
-    (100.0, 101.0),
-    (101.0, 102.0),
-    (102.0, 103.0),
-    (102.0, 148.0),
-    (103.0, 104.0),
-    (104.0, 105.0),
-    (104.0, 138.0),
-    (105.0, 106.0),
-    (138.0, 139.0),
-    (139.0, 140.0),
-    (140.0, 141.0),
-    (141.0, 142.0),
-    (142.0, 143.0),
-    (143.0, 144.0),
-    (144.0, 145.0),
-    (145.0, 146.0),
-    (146.0, 147.0),
-    (148.0, 149.0),
-    (149.0, 150.0),
-]
-
-small_tree_edges = [
-    (0.0, -1.0),
-    (0.0, 1.0),
-    (1.0, 2.0),
-    (2.0, 3.0),
-    (3.0, 4.0),
-    (4.0, 5.0),
-    (5.0, 6.0),
-    (6.0, 7.0),
-    (7.0, 8.0),
-    (8.0, 9.0),
-    (9.0, 10.0),
-    (10.0, 11.0),
-    (11.0, 12.0),
-    (12.0, 13.0),
-    (13.0, 14.0),
-    (14.0, 15.0),
-    (15.0, 16.0),
-    (16.0, 17.0),
-    (17.0, 18.0),
-    (18.0, 19.0),
-    (19.0, 20.0),
-    (20.0, 21.0),
-    (21.0, 22.0),
-    (22.0, 23.0),
-    (23.0, 24.0),
-    (24.0, 25.0),
-    (25.0, 26.0),
-    (26.0, 27.0),
-    (27.0, 28.0),
-    (28.0, 29.0),
-    (29.0, 30.0),
-    (30.0, 31.0),
-    (31.0, 32.0),
-    (32.0, 33.0),
-    (33.0, 34.0),
-    (34.0, 35.0),
-    (35.0, 36.0),
-    (36.0, 37.0),
-    (37.0, 38.0),
-    (38.0, 39.0),
-    (39.0, 40.0),
-    (40.0, 41.0),
-    (41.0, 42.0),
-    (42.0, 43.0),
-    (43.0, 44.0),
-    (44.0, 45.0),
-    (45.0, 46.0),
-    (46.0, 47.0),
-    (47.0, 48.0),
-    (48.0, 49.0),
-    (49.0, 50.0),
-    (50.0, 51.0),
-    (51.0, 52.0),
-    (52.0, 53.0),
-    (53.0, 54.0),
-    (54.0, 55.0),
-    (55.0, 56.0),
-    (56.0, 57.0),
-    (57.0, 58.0),
-    (58.0, 59.0),
-    (59.0, 60.0),
-    (60.0, 61.0),
-    (61.0, 62.0),
-    (62.0, 63.0),
-    (63.0, 64.0),
-    (64.0, 65.0),
-    (65.0, 66.0),
-    (66.0, 67.0),
-    (67.0, 68.0),
-    (68.0, 69.0),
-    (69.0, 70.0),
-    (70.0, 71.0),
-    (71.0, 72.0),
-    (72.0, 73.0),
-    (73.0, 74.0),
-    (74.0, 75.0),
-    (75.0, 76.0),
-    (76.0, 77.0),
-    (77.0, 78.0),
-    (78.0, 79.0),
-    (79.0, 80.0),
-    (80.0, 81.0),
-    (81.0, 82.0),
-    (82.0, 83.0),
-    (83.0, 84.0),
-    (84.0, 85.0),
-    (85.0, 86.0),
-    (86.0, 87.0),
-    (87.0, 88.0),
-    (88.0, 89.0),
-    (89.0, 90.0),
-    (90.0, 91.0),
-    (91.0, 92.0),
-    (92.0, 93.0),
-    (93.0, 94.0),
-    (94.0, 95.0),
-    (95.0, 96.0),
-    (96.0, 97.0),
-    (97.0, 98.0),
-    (98.0, 99.0),
-    (99.0, 100.0),
-    (100.0, 101.0),
-    (101.0, 102.0),
-    (102.0, 103.0),
-    (103.0, 104.0),
-    (104.0, 105.0),
-    (105.0, 106.0),
-    (106.0, 107.0),
-    (107.0, 108.0),
-    (108.0, 109.0),
-    (109.0, 110.0),
-    (110.0, 111.0),
-    (111.0, 112.0),
-    (112.0, 113.0),
-    (113.0, 114.0),
-    (114.0, 115.0),
-    (115.0, 116.0),
-    (116.0, 117.0),
-    (117.0, 118.0),
-    (117.0, 412.0),
-    (118.0, 119.0),
-    (119.0, 120.0),
-    (120.0, 121.0),
-    (121.0, 122.0),
-    (122.0, 123.0),
-    (122.0, 353.0),
-    (123.0, 124.0),
-    (124.0, 125.0),
-    (125.0, 126.0),
-    (126.0, 127.0),
-    (126.0, 277.0),
-    (127.0, 128.0),
-    (128.0, 129.0),
-    (129.0, 130.0),
-    (130.0, 131.0),
-    (130.0, 228.0),
-    (131.0, 132.0),
-    (132.0, 133.0),
-    (133.0, 134.0),
-    (133.0, 224.0),
-    (133.0, 226.0),
-    (134.0, 135.0),
-    (134.0, 192.0),
-    (135.0, 136.0),
-    (136.0, 137.0),
-    (137.0, 138.0),
-    (138.0, 139.0),
-    (138.0, 190.0),
-    (139.0, 140.0),
-    (139.0, 165.0),
-    (139.0, 186.0),
-    (140.0, 141.0),
-    (140.0, 155.0),
-    (141.0, 142.0),
-    (142.0, 143.0),
-    (143.0, 144.0),
-    (143.0, 152.0),
-    (144.0, 145.0),
-    (145.0, 146.0),
-    (145.0, 149.0),
-    (145.0, 151.0),
-    (146.0, 147.0),
-    (147.0, 148.0),
-    (149.0, 150.0),
-    (152.0, 153.0),
-    (153.0, 154.0),
-    (155.0, 156.0),
-    (156.0, 157.0),
-    (157.0, 158.0),
-    (158.0, 159.0),
-    (159.0, 160.0),
-    (160.0, 161.0),
-    (161.0, 162.0),
-    (162.0, 163.0),
-    (163.0, 164.0),
-    (165.0, 166.0),
-    (165.0, 177.0),
-    (165.0, 184.0),
-    (166.0, 167.0),
-    (167.0, 168.0),
-    (168.0, 169.0),
-    (168.0, 174.0),
-    (169.0, 170.0),
-    (169.0, 172.0),
-    (170.0, 171.0),
-    (172.0, 173.0),
-    (174.0, 175.0),
-    (175.0, 176.0),
-    (177.0, 178.0),
-    (178.0, 179.0),
-    (179.0, 180.0),
-    (180.0, 181.0),
-    (181.0, 182.0),
-    (182.0, 183.0),
-    (184.0, 185.0),
-    (186.0, 187.0),
-    (187.0, 188.0),
-    (188.0, 189.0),
-    (190.0, 191.0),
-    (192.0, 193.0),
-    (192.0, 222.0),
-    (193.0, 194.0),
-    (194.0, 195.0),
-    (195.0, 196.0),
-    (196.0, 197.0),
-    (196.0, 220.0),
-    (197.0, 198.0),
-    (198.0, 199.0),
-    (199.0, 200.0),
-    (199.0, 214.0),
-    (200.0, 201.0),
-    (201.0, 202.0),
-    (201.0, 212.0),
-    (202.0, 203.0),
-    (203.0, 204.0),
-    (204.0, 205.0),
-    (204.0, 209.0),
-    (205.0, 206.0),
-    (206.0, 207.0),
-    (207.0, 208.0),
-    (209.0, 210.0),
-    (210.0, 211.0),
-    (212.0, 213.0),
-    (214.0, 215.0),
-    (215.0, 216.0),
-    (215.0, 218.0),
-    (216.0, 217.0),
-    (218.0, 219.0),
-    (220.0, 221.0),
-    (222.0, 223.0),
-    (224.0, 225.0),
-    (226.0, 227.0),
-    (228.0, 229.0),
-    (229.0, 230.0),
-    (229.0, 267.0),
-    (230.0, 231.0),
-    (231.0, 232.0),
-    (232.0, 233.0),
-    (233.0, 234.0),
-    (234.0, 235.0),
-    (235.0, 236.0),
-    (235.0, 256.0),
-    (235.0, 263.0),
-    (236.0, 237.0),
-    (236.0, 253.0),
-    (237.0, 238.0),
-    (238.0, 239.0),
-    (239.0, 240.0),
-    (240.0, 241.0),
-    (240.0, 251.0),
-    (241.0, 242.0),
-    (242.0, 243.0),
-    (242.0, 249.0),
-    (243.0, 244.0),
-    (244.0, 245.0),
-    (245.0, 246.0),
-    (245.0, 248.0),
-    (246.0, 247.0),
-    (249.0, 250.0),
-    (251.0, 252.0),
-    (253.0, 254.0),
-    (254.0, 255.0),
-    (256.0, 257.0),
-    (257.0, 258.0),
-    (257.0, 261.0),
-    (258.0, 259.0),
-    (259.0, 260.0),
-    (261.0, 262.0),
-    (263.0, 264.0),
-    (264.0, 265.0),
-    (265.0, 266.0),
-    (267.0, 268.0),
-    (268.0, 269.0),
-    (269.0, 270.0),
-    (270.0, 271.0),
-    (271.0, 272.0),
-    (272.0, 273.0),
-    (273.0, 274.0),
-    (274.0, 275.0),
-    (275.0, 276.0),
-    (277.0, 278.0),
-    (278.0, 279.0),
-    (279.0, 280.0),
-    (280.0, 281.0),
-    (281.0, 282.0),
-    (282.0, 283.0),
-    (283.0, 284.0),
-    (284.0, 285.0),
-    (284.0, 328.0),
-    (284.0, 352.0),
-    (285.0, 286.0),
-    (285.0, 323.0),
-    (286.0, 287.0),
-    (287.0, 288.0),
-    (287.0, 312.0),
-    (288.0, 289.0),
-    (288.0, 300.0),
-    (288.0, 307.0),
-    (289.0, 290.0),
-    (290.0, 291.0),
-    (291.0, 292.0),
-    (291.0, 298.0),
-    (292.0, 293.0),
-    (293.0, 294.0),
-    (294.0, 295.0),
-    (295.0, 296.0),
-    (296.0, 297.0),
-    (298.0, 299.0),
-    (300.0, 301.0),
-    (301.0, 302.0),
-    (302.0, 303.0),
-    (302.0, 305.0),
-    (303.0, 304.0),
-    (305.0, 306.0),
-    (307.0, 308.0),
-    (308.0, 309.0),
-    (309.0, 310.0),
-    (310.0, 311.0),
-    (312.0, 313.0),
-    (313.0, 314.0),
-    (313.0, 318.0),
-    (313.0, 321.0),
-    (314.0, 315.0),
-    (315.0, 316.0),
-    (316.0, 317.0),
-    (318.0, 319.0),
-    (319.0, 320.0),
-    (321.0, 322.0),
-    (323.0, 324.0),
-    (324.0, 325.0),
-    (325.0, 326.0),
-    (325.0, 327.0),
-    (328.0, 329.0),
-    (329.0, 330.0),
-    (330.0, 331.0),
-    (330.0, 346.0),
-    (330.0, 350.0),
-    (331.0, 332.0),
-    (332.0, 333.0),
-    (332.0, 344.0),
-    (333.0, 334.0),
-    (334.0, 335.0),
-    (334.0, 339.0),
-    (334.0, 342.0),
-    (335.0, 336.0),
-    (335.0, 338.0),
-    (336.0, 337.0),
-    (339.0, 340.0),
-    (340.0, 341.0),
-    (342.0, 343.0),
-    (344.0, 345.0),
-    (346.0, 347.0),
-    (347.0, 348.0),
-    (348.0, 349.0),
-    (350.0, 351.0),
-    (353.0, 354.0),
-    (354.0, 355.0),
-    (355.0, 356.0),
-    (356.0, 357.0),
-    (356.0, 389.0),
-    (356.0, 410.0),
-    (357.0, 358.0),
-    (358.0, 359.0),
-    (359.0, 360.0),
-    (359.0, 384.0),
-    (360.0, 361.0),
-    (360.0, 373.0),
-    (361.0, 362.0),
-    (362.0, 363.0),
-    (363.0, 364.0),
-    (363.0, 371.0),
-    (364.0, 365.0),
-    (364.0, 367.0),
-    (364.0, 369.0),
-    (365.0, 366.0),
-    (367.0, 368.0),
-    (369.0, 370.0),
-    (371.0, 372.0),
-    (373.0, 374.0),
-    (374.0, 375.0),
-    (375.0, 376.0),
-    (375.0, 382.0),
-    (376.0, 377.0),
-    (377.0, 378.0),
-    (378.0, 379.0),
-    (378.0, 381.0),
-    (379.0, 380.0),
-    (382.0, 383.0),
-    (384.0, 385.0),
-    (385.0, 386.0),
-    (386.0, 387.0),
-    (387.0, 388.0),
-    (389.0, 390.0),
-    (390.0, 391.0),
-    (391.0, 392.0),
-    (392.0, 393.0),
-    (393.0, 394.0),
-    (393.0, 408.0),
-    (394.0, 395.0),
-    (395.0, 396.0),
-    (395.0, 403.0),
-    (396.0, 397.0),
-    (396.0, 401.0),
-    (397.0, 398.0),
-    (398.0, 399.0),
-    (399.0, 400.0),
-    (401.0, 402.0),
-    (403.0, 404.0),
-    (403.0, 407.0),
-    (404.0, 405.0),
-    (405.0, 406.0),
-    (408.0, 409.0),
-    (410.0, 411.0),
-    (412.0, 413.0),
-    (413.0, 414.0),
-    (414.0, 415.0),
-    (415.0, 416.0),
-    (416.0, 417.0),
-    (417.0, 418.0),
-    (418.0, 419.0),
-    (418.0, 490.0),
-    (419.0, 420.0),
-    (420.0, 421.0),
-    (421.0, 422.0),
-    (422.0, 423.0),
-    (423.0, 424.0),
-    (423.0, 470.0),
-    (423.0, 487.0),
-    (424.0, 425.0),
-    (424.0, 460.0),
-    (425.0, 426.0),
-    (426.0, 427.0),
-    (427.0, 428.0),
-    (428.0, 429.0),
-    (429.0, 430.0),
-    (430.0, 431.0),
-    (430.0, 456.0),
-    (431.0, 432.0),
-    (431.0, 454.0),
-    (432.0, 433.0),
-    (433.0, 434.0),
-    (434.0, 435.0),
-    (434.0, 453.0),
-    (435.0, 436.0),
-    (435.0, 451.0),
-    (436.0, 437.0),
-    (437.0, 438.0),
-    (438.0, 439.0),
-    (439.0, 440.0),
-    (440.0, 441.0),
-    (441.0, 442.0),
-    (442.0, 443.0),
-    (443.0, 444.0),
-    (444.0, 445.0),
-    (445.0, 446.0),
-    (445.0, 449.0),
-    (446.0, 447.0),
-    (447.0, 448.0),
-    (449.0, 450.0),
-    (451.0, 452.0),
-    (454.0, 455.0),
-    (456.0, 457.0),
-    (457.0, 458.0),
-    (458.0, 459.0),
-    (460.0, 461.0),
-    (461.0, 462.0),
-    (462.0, 463.0),
-    (463.0, 464.0),
-    (463.0, 468.0),
-    (464.0, 465.0),
-    (465.0, 466.0),
-    (466.0, 467.0),
-    (468.0, 469.0),
-    (470.0, 471.0),
-    (470.0, 480.0),
-    (471.0, 472.0),
-    (471.0, 475.0),
-    (471.0, 478.0),
-    (472.0, 473.0),
-    (473.0, 474.0),
-    (475.0, 476.0),
-    (476.0, 477.0),
-    (478.0, 479.0),
-    (480.0, 481.0),
-    (481.0, 482.0),
-    (482.0, 483.0),
-    (483.0, 484.0),
-    (484.0, 485.0),
-    (485.0, 486.0),
-    (487.0, 488.0),
-    (488.0, 489.0),
-    (490.0, 491.0),
-    (491.0, 492.0),
-    (492.0, 493.0),
-    (492.0, 516.0),
-    (493.0, 494.0),
-    (493.0, 508.0),
-    (494.0, 495.0),
-    (495.0, 496.0),
-    (495.0, 504.0),
-    (496.0, 497.0),
-    (497.0, 498.0),
-    (498.0, 499.0),
-    (498.0, 501.0),
-    (498.0, 503.0),
-    (499.0, 500.0),
-    (501.0, 502.0),
-    (504.0, 505.0),
-    (505.0, 506.0),
-    (506.0, 507.0),
-    (508.0, 509.0),
-    (509.0, 510.0),
-    (509.0, 513.0),
-    (509.0, 515.0),
-    (510.0, 511.0),
-    (511.0, 512.0),
-    (513.0, 514.0),
-]
-drip_on_trunk_flows = [
-    Flow(
-        num_cylinders=100,
-        projected_area=18.585135753681307,
-        surface_area=22.124462796803396,
-        angle_sum=121.15837282927401,
-        volume=4.300205,
-        sa_to_vol=1486.9747724424142,
-        drip_node_id=0,
-        drip_node_loc=(0.04115, -3.741772),
-    ),
-    Flow(
-        num_cylinders=98,
-        projected_area=0.5738643958750281,
-        surface_area=1.4981470618875297,
-        angle_sum=39.99693438992534,
-        volume=0.046667,
-        sa_to_vol=51223.61034123854,
-        drip_node_id=100.0,
-        drip_node_loc=(-2.378381, -2.907627, 12.366749),
-    ),
-]
-
-ten_cyls_dbh = 0.37489
-happy_path_dbh = 0.37489
-small_tree_dbh = 1.058276
