--- src/canhydro/tests/test_geometry_integration.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-
-def test_lapack_v_lu_factor():
-    """
-    Tests thse answers are the same
-    will be easy, quick and minory beneficial
-    """
