--- src/canhydro/tests/test_utils.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-
-def test_stack():
-    """
-    Test vstack and col stack capabilities v. numpy
-    thow in a couple exenmplar custom tests to show functionality
-    """
