+++ test/expected_results.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from canhydro.DataClasses import Flow
+from src.canhydro.DataClasses import Flow
 
 ten_cyls_cyls = [
     "Cylinder(cyl_id=0.0, x=array([0.076822, 0.076822]), y=array([-3.386935, -3.386935]), z=array([-0.595933, -0.5315  ]), radius=0.222504, length=0.064433, branch_order=0.0, branch_id=0.0, volume=0.010021, parent_id=-1.0, reverse_branch_order=32.0, segment_id=0.0, projected_data={}, flow_id=None, flow_type=None, begins_at_drip_point=None, begins_at_divide_point=None, dx=0.0, dy=0.0, dz=0.06443300000000007, surface_area=0.09007951593260982, sa_to_vol=8.98907453673384, slope=0.0, is_stem=False, is_divide=False)",
@@ -798,6 +798,460 @@ small_tree_is_stem = (
     False,
 )
 
+drip_mid_stem_map = (
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+)
+
+drip_adj_stem_map = (
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+)
+
+drip_on_trunk_stem_map = (
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    True,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+    False,
+)
 
 ten_cyls_rows_proj_XZ = []
 ten_cyls_rows_proj_XY = []
@@ -1019,6 +1473,52 @@ small_tree_flows = [
         sa_to_vol=81927.57367017229,
         drip_node_id=513.0,
         drip_node_loc=(1.581945, 2.857679, 12.955942),
+    ),
+]
+
+drip_mid_flows = [
+    Flow(
+        num_cylinders=117,
+        projected_area=16.47655636448612,
+        surface_area=19.350321212472778,
+        angle_sum=134.88157931803016,
+        volume=3.906033,
+        sa_to_vol=7306.133204864248,
+        drip_node_id=0,
+        drip_node_loc=(-0.299115, 2.537844),
+    ),
+    Flow(
+        num_cylinders=14,
+        projected_area=0.010439067766281047,
+        surface_area=0.03744675454760344,
+        angle_sum=8.673607609925796,
+        volume=0.000121,
+        sa_to_vol=4804.1689124433915,
+        drip_node_id=117.0,
+        drip_node_loc=(1.011551, 2.485898, 13.266762),
+    ),
+]
+
+drip_adj_flows = [
+    Flow(
+        num_cylinders=107,
+        projected_area=2.9604039491372682,
+        surface_area=7.394338347069087,
+        angle_sum=124.04233409518383,
+        volume=0.521757,
+        sa_to_vol=4664.192277179154,
+        drip_node_id=0,
+        drip_node_loc=(0.076822, -3.386935),
+    ),
+    Flow(
+        num_cylinders=10,
+        projected_area=0.006539752252989748,
+        surface_area=0.027999615993246064,
+        angle_sum=7.133359698969585,
+        volume=5.2e-05,
+        sa_to_vol=6917.96517821172,
+        drip_node_id=125.0,
+        drip_node_loc=(-0.170586, -1.090298, 10.411159),
     ),
 ]
 
@@ -1677,7 +2177,28 @@ small_tree_edges = [
     (511.0, 512.0),
     (513.0, 514.0),
 ]
-
+drip_on_trunk_flows = [
+    Flow(
+        num_cylinders=100,
+        projected_area=18.585135753681307,
+        surface_area=22.124462796803396,
+        angle_sum=121.15837282927401,
+        volume=4.300205,
+        sa_to_vol=1486.9747724424142,
+        drip_node_id=0,
+        drip_node_loc=(0.04115, -3.741772),
+    ),
+    Flow(
+        num_cylinders=98,
+        projected_area=0.5738643958750281,
+        surface_area=1.4981470618875297,
+        angle_sum=39.99693438992534,
+        volume=0.046667,
+        sa_to_vol=51223.61034123854,
+        drip_node_id=100.0,
+        drip_node_loc=(-2.378381, -2.907627, 12.366749),
+    ),
+]
 
 ten_cyls_dbh = 0.37489
 happy_path_dbh = 0.37489
