--- user_def_config.toml
@@ -1,52 +1,0 @@
-
-[qsm]
-# type=0
-cyl_id=1
-parent_id=2
-x=[3,6]
-y=[4,7]
-z=[5,8]
-radius=9
-volume=10
-# growth_volume=11
-length=12
-# growth_length=13
-# fitting_type=14
-segment_id=15
-# parent_segment_id=16
-# segment_median_radius=17
-# segment_growth_volume=18
-# segment_growth_length=19
-branch_order=20
-reverse_branch_order=21
-# reverse_pipe_radius_branch_order=22
-# reverse_pipe_area_branch_order=23
-branch_id=24
-# tree_id=25
-# tree_species=26
-# translate_xyz=[27,28,29]
-# gv= [30,31,32]
-# gl= [33,34]
-
-[project.optional-dependencies]
-test = [
-    "black==23.3.0",
-    "check-manifest==0.49",
-    "pre-commit==3.3.1",
-    "pylint==2.17.4",
-    "pylint_junit",
-    "pytest-cov==4.0.0",
-    "pytest-mock<3.10.1",
-    "pytest-runner",
-    "pytest==7.3.1",
-    "pytest-github-actions-annotate-failures",
-    "shellcheck-py==0.9.0.2",
-]
-
-[project.urls]
-Documentation = "https://github.com/wischmcj/canopyHydrodynamics/blob/main/README.md"
-Source = "https://github.com/wischmcj/canopyHydrodynamics/"
-
-
-[directories]
-root = "C:/Users/wisch/Documents/GitProjects/CanopyHydrodynamics/"
